+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Enh #282: Support `numeric` arrays, improve support of domain types and `int` and `varchar` array types (@Tigrov)
 - Enh #284: Add tests for `binary` type and fix casting of default value (@Tigrov)
 - Chg #288: Typecast refactoring (@Tigrov)
+- Bug #287: Fix `bit` type (@Tigrov)
+- Enh #289: Array parser refactoring (@Tigrov)
 
 ## 1.0.0 April 12, 2023